src/backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-dotenv==0.19.0
celery==5.1.2
redis==3.5.3
elasticsearch==7.14.0